use appropriate CA cert to reach the front-proxy
The kcp chart allows passing a custom kcpFrontProxy.certificateIssuer
to provide the front-proxy a server cert, instead of the default
kcp-server-issuer. However, other parts of the chart (notably the
front-proxy's ServiceMonitor and the external-admin-kubeconfig) were not
properly reacting to this setting.

More precisely, here is the CA cert that should be used by components
that try to reach the kcp-front-proxy:
- when no certificateIssuer is provided: the default kcp-ca
- when a *public* certificateIssuer is provided (e.g. LetsEncrypt):
  no CA
- when a custom *private* certificateIssuser is provided: the CA cert
  for that private issuer, referenced via kcpFrontProxy.certificateIssuer.secret

Signed-off-by: Erwan DE LEPINAU <elepinau@evroc.com>