remove dedicated options to define RBAC
This offered no real benefit over someone simply having their own
rbac.yaml. Adding this as an option to the chart just means we
have to continue to support it, and why would we then only support
defining custom RBAC and not even other resources? At that point we
would become a meta-meta Helm chart and that's just not worth it.

On-behalf-of: @SAP christoph.mewes@sap.com